Initial StyleRule implementation
